src/twilio/requirements.txt
twilio
python-dotenv
openai